requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
cohere==5.13.12
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
django-ninja==1.3.0
django-use-email-as-username==1.4.0
fastavro==1.10.0
filelock==3.17.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
idna==3.10
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
tokenizers==0.21.0
tqdm==4.67.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.9.0
